Implement maze size prompt and DFS changes
Prompt for desired maze size added and implemented. Changes to the directions for the adjacency features needed to be implemented to accommodate this. Changes to DFS were also made, mainly adding end cases to ensure the algorithm stops when it reaches the end tile. General code cleanup was implemented as well.