Implement various clearing, generation, and searching functions
This is a quite substantial commit that adds in numerous functionalities. One such functionality is the ability to clear walls and reset algorithms. Another would be the implementation of automatic maze generation that can be performed as much as the user prefers during runtime. Finally, the implementation of a random search was added as an algorithm choice.